backend/requirements.txt
# ── RAG pipeline dependencies ──
langchain-classic
langchain-community
langchain-core
langchain-huggingface
langchain-text-splitters
langchain-experimental
langchain-chroma
chromadb
sentence-transformers
PyMuPDF
rank-bm25
python-dotenv
langchain-ollama

# ── Backend server ──
fastapi
uvicorn[standard]
